Add Hibernate dependencies and implement functionality for managing Tickets and Users. Implement features to: - Save Tickets and Users. - Fetch Tickets by ID and user ID, and Users by ID. - Update ticket types. - Delete Users by ID and cascade delete their tickets if any. - Add functionality to update both a User and all their tickets within the same transaction using Hibernate.

+++ andersen-homework/src/main/java/org/example/dao/TicketDao.java
@@ -1,7 +1,8 @@
 package org.example.dao;
 
+
+import org.example.entity.Ticket;
 import org.example.enums.TicketType;
-import org.example.model.Ticket;
 
 import java.sql.SQLException;
 import java.util.List;
@@ -12,7 +13,7 @@ public interface TicketDao {
 
     List<Ticket> getTicketsByUserId(int userId);
 
-    Ticket getTicketById(int id) throws SQLException;
+    Ticket getTicketById(Long id) throws SQLException;
 
     void updateTicketType(int ticketId, TicketType ticketType) throws SQLException;
